requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
twilio==8.12.0
python-dotenv==1.0.0
pandas==2.1.4
pydantic==2.5.2
vobject==0.9.6.1
pydantic[email]==2.5.2
requests==2.31.0
africastalking==1.2.9

